updated end screen and infinite loop is now functional
